Fix the Icon class not supporting colors with alpha
ColorOverlay simply "overlays" the color on the image, so a color with alpha means that instead of a partially transparent icon, you get an icon that is a mix of black and the requested color.

This solves that by changing the overlay color to 100% opacity and then changing the opacity of the overlay iteself.