[Snackbar] Fix button margins and touch target size
Since the snackbar button margin in Material Design spec
is measured to the button label border, not to the button
border, I've added the new possible context for Button type
that doesn't limit the minimal button size and fixed the
margins to position button correctly when shown in toast.

Also needed to enhance button height a bit to comply with
Material Design minimum button height of 36dp.