Fix issues on mobile/touch devices
Touch devices do not interact very well with the "hover" property of the mouse area. This causes the widgets that use this object to get "stuck" with the color used when the item is clicked, even if the user is not touching/clicking the widget.

To fix this, we simply disable the hover feature when running on a mobile device.