Suppress meaningless error on ProgressCircle
In some cases where ProgressCircle's dimensions are set through property bindings, when the app is initially starting up and the bindings haven't resolved yet, the ProgressCircle's dimensions will be zero. This causes a negative radius when drawing the arc, which prints an error to the console. The behavior is not affected; once the bindings resolve and the ProgressCircle has non-zero dimensions, it works normally. Suppress this error by lower-bounding the radius to zero, as nothing has actually gone wrong.