Allow the user not colorize a ListItem.Standard element
This is a very simple addition for development of the papyros/files-app.
It allows a user to choose whether the icon on a ListItem.Standard
is colorized or not. In a ListItem.Standard a user can now use the
field "colorizeIcon:" which is an alias for "icon.colorize" in the
item.

This came to be an issue when attempting to list mimetype icons
in the papyros files app as they were all using a ColorOverlay instead
of an image.

Changes:
	1. Added a colorizeIcon property in ListItem.Standard
	   which allows a user to turn the ColorOverlay on and
	   off at will.