Revert "Merge pull request #1 from Spateof-io/QMLMAT-1-enable-switch-insert-in-ActionBar"
This reverts commit 12c45a7dbf14405e92662d102b0d3a5b21b64e02, reversing
changes made to 0cada50ee5469efdeb37e9eb2f5e8de0194b6ba6.

As we can use directly the customContent property of the actionBar, no
need to provide such feature and properties to add switches to the
actionbar. the icon bug correction will be restaured in a standalone
branch for a separated pull requests.